Migrated make_bip32_path from rust-app to the SDK
The make_bip32_path macro has been added in rust-app by pscott. It is
even better to have it in the SDK.
The macro has also been converted to a const fn. There is no particular
need for it to be a macro, and a const fn will provide the guaranty that
it is evaluated at compile time.