Use #38 in build.rs and structure
The first change is to parse once, and then use an enum for the device.
This is hopefully a straightforward improvement.

The second change is to case on the OS name (leveraging #38) instead of
the target name. The target "name" isn't so structured, and it is
unclear to what extent it should be anm exposed part of the target. (See
https://github.com/rust-lang/rust/pull/98225 for example, where the
contents rather than json file path were used as a a key.)

With #38 the device name is used for the OS field instead, and so we are
robust to confusing behavior around names.